requirements.txt
langchain
langchain-core


langchain-google-genai
google-generativeai
python-dotenv
numpy
youtube-transcript-api
langchain-community
faiss-cpu 
tiktoken
streamlit